requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
pydantic>=2.0.0
google-generativeai>=0.8.5
